Move valves to separate package
